fix: keep doctor formatting within package boundary
